move span into data utils
