+++ lib/data/src/lib.rs
@@ -3,5 +3,6 @@
 #![warn(missing_debug_implementations)]
 
 pub mod opt_vec;
+pub mod text;
 
 pub use enso_prelude as prelude;
