+++ lib/ide/ast/impl/Cargo.toml
@@ -18,3 +18,4 @@ uuid         = { version  = "0.8.1", features = ["serde", "v4"] }
 ast-macros   = { version = "0.1.0", path = "../macros"             }
 enso-prelude = { version = "0.1.0", path = "../../../prelude"      }
 shapely      = { version = "0.1.0", path = "../../../shapely/impl" }
+data         = { version = "0.1.0", path = "../../../data"         }
