+++ lib/ide/parser/src/main.rs
@@ -1,6 +1,7 @@
+use enso_prelude::*;
+
 use parser::api::IsParser;
 
-use enso_prelude::default;
 
 /// Simple interactive tester - calls parser with its argument (or a
 /// hardcoded default) and prints the result
