+++ lib/ide/parser/src/wsclient.rs
@@ -9,6 +9,8 @@ use websocket::{
 use crate::api;
 
 use api::Error::*;
+use ast::IDMap;
+
 use Error::*;
 
 type WsTcpClient = websocket::sync::Client<TcpStream>;
@@ -85,7 +87,7 @@ impl From<serde_json::error::Error> for Error {
 /// All request supported by the Parser Service.
 #[derive(Debug, serde::Serialize, serde::Deserialize)]
 pub enum Request {
-    ParseRequest { program: String, ids: api::IDMap },
+    ParseRequest { program:String, ids:IDMap },
 }
 
 /// All responses that Parser Service might reply with.
@@ -198,8 +200,8 @@ impl Client {
 
 impl api::IsParser for Client {
 
-   fn parse(&mut self, program:String, ids:api::IDMap) -> api::Result<api::Ast> {
-        let request  = Request::ParseRequest { program, ids };
+   fn parse(&mut self, program:String, ids:IDMap) -> api::Result<api::Ast> {
+        let request  = Request::ParseRequest {program,ids};
         let response = self.rpc_call(request)?;
         match response {
             Response::Success { ast_json } => internal::from_json(&ast_json),
